+++ lib/ivcv/result.ex
@@ -10,6 +10,8 @@ defmodule IvcvEx.Result.Impression do
         }
   @enforce_keys [:neuroticism, :extraversion, :openness, :conscientiousness, :agreeableness]
 
+  alias IvcvEx.Result.Impression
+
   defstruct [
     :neuroticism,
     :extraversion,
@@ -25,7 +27,7 @@ defmodule IvcvEx.Result.Impression do
         "conscientiousness" => conscientiousness,
         "agreeableness" => agreeableness
       }) do
-    %{
+    %Impression{
       neuroticism: neuroticism,
       extraversion: extraversion,
       openness: openness,
@@ -43,22 +45,35 @@ defmodule IvcvEx.Result.Emotions do
           positivity: integer()
         }
   @enforce_keys [:intensity, :positivity]
+
+  alias IvcvEx.Result.Emotions
+
   defstruct [
     :intensity,
     :positivity
   ]
 
   def parse(%{"intensity" => intensity, "positivity" => positivity}) do
-    %{intensity: intensity, positivity: positivity}
+    %Emotions{intensity: intensity, positivity: positivity}
   end
 end
 
 defmodule IvcvEx.Result do
   @moduledoc """
   Struct to hold the data analysis result
+    emotions (Emotions.t()) - Average intensity and positivity during the video. Integer values between 0 to 100.
+
+    impression (Impression.t()) - First impression analysis with Big Five metrics. Integer values between 0 to 100.
+
+    sentiment (float) - Sentiment analysis result. Integer value between 0 to 100.
+
+    status (string) - Current status of the processing. One of FINISHED, PROCESSING, FAILED.
+
+    resultId (string) - Id of the result.
   """
   require Logger
 
+  alias IvcvEx.Result
   alias IvcvEx.Result.Emotions
   alias IvcvEx.Result.Impression
 
@@ -68,7 +83,7 @@ defmodule IvcvEx.Result do
           status: String.t(),
           impression: Impression,
           emotions: Emotions,
-          sentiment: integer()
+          sentiment: float()
         }
   defstruct [
     :result_id,
@@ -78,18 +93,6 @@ defmodule IvcvEx.Result do
     :sentiment
   ]
 
-  def fragment() do
-    """
-    {
-      result_id
-      status
-      impression
-      emotions
-      sentiment
-    }
-    """
-  end
-
   def parse(%{
         "resultId" => result_id,
         "status" => status,
@@ -98,12 +101,21 @@ defmodule IvcvEx.Result do
         "sentiment" => sentiment
       }) do
     {:ok,
-     %{
+     %Result{
        result_id: result_id,
        status: status,
        impression: Impression.parse(impression),
        emotions: Emotions.parse(emotions),
        sentiment: sentiment
+     }}
+  end
+
+  def parse(%{
+        "resultId" => result_id
+      }) do
+    {:ok,
+     %{
+       result_id: result_id
      }}
   end