+++ lib/ivcv_ex.ex
@@ -46,6 +46,31 @@ defmodule IvcvEx do
     |> parse_response()
   end
 
+  @doc """
+  Get the result of a video analysis. Result include multiple items from the analysis.
+
+  Returns `{:ok, Result.t()}` and Result contains:
+    emotions (object) - Average intensity and positivity during the video. Integer values between 0 to 100.
+
+    impression (object) - First impression analysis with Big Five metrics. Integer values between 0 to 100.
+
+    sentiment (float) - Sentiment analysis result. Integer value between 0 to 100.
+
+    status (string) - Current status of the processing. One of FINISHED, PROCESSING, FAILED.
+
+    resultId (string) - Id of the result.
+
+  ## Parameters
+
+    - result_id: Id of the result. This is obtained when the video is sent initially.
+
+  ## Examples
+
+      IvcvEx.get_analysis_result("result_id")
+
+  ## Reference
+    https://docs.ivcv.eu/result_for_video.html#blueprints.direct_blueprints.analysis_result
+  """
   def get_analysis_result(result_id) do
     path = "/result"
     auth_api_key = Application.get_env(:ivcv_ex, :auth_key)